#3 correccion de estructura
se corrigio la estructura del proyecto
